backend/requirements.txt
django>=3.1
djangorestframework>=3.12
django-webpack-loader==0.7.0
django-cors-headers>=3.7